Add a <elementary-math> math custom element so elementary math works outside of MathML context. This is for experimention.
Changed the way the polyfill modifies the DOM. It is now, IMHO, cleaner.
The polyfill adds a span around the math element and puts a shadow root on the span.
Hence, the DOM still has unmodified MathML in it.

A question to be answered is how other polyfills interact with this.
A quick test found that another polyfill could put a shadowRoot inside the shadowRoot created here, but the polyfill would need to know to modify the shadowRoot and not the original MathML.